__meta/synchronizer/requirements.txt
pygithub
python-dotenv
python-keycloak
hvac
slack_sdk
